+++ config/foot_config.go
@@ -1,0 +1,6 @@
+package config
+
+import _ "embed"
+
+//go:embed foot.ini
+var FootConfig string